fix: fix beberapa bug

+++ app/Http/Controllers/Api/ManajemenKotaController.php
@@ -9,33 +9,37 @@ use Illuminate\Http\Request;
 class ManajemenKotaController extends Controller
 {
 
-    public function index(){
+    public function index()
+    {
         $kota = Kota::all();
-        return response()->json(['kota'=>$kota], 200);
+        return response()->json(['kota' => $kota], 200);
     }
 
-    public function tambahKota(Request $request){
+    public function tambahKota(Request $request)
+    {
 
         $request->validate([
-            'kota' => 'required',
+            'kota' => 'required|unique:kotas,kota',
         ]);
 
         $kota = Kota::create([
-            'kota' => $request->kota,
-        ]); 
+            'kota' => $request->kota
+        ]);
 
-        return response()->json(['kota'=>$kota], 200);
+        return response()->json(['message' => 'Berhasil Menambahkan Kota',], 201);
     }
 
-    public function editKota($id){
+    public function editKota($id)
+    {
 
         $kota = Kota::find($id);
 
-        return response()->json(['kota'=>$kota], 200);
+        return response()->json(['kota' => $kota], 200);
     }
 
 
-    public function updateKota(Request $request, $id){
+    public function updateKota(Request $request, $id)
+    {
         $request->validate([
             'kota' => 'required',
         ]);
@@ -44,13 +48,21 @@ class ManajemenKotaController extends Controller
         $kota->kota = $request->kota;
         $kota->save();
 
-        return response()->json(['message'=>'succes update'], 200);
+        return response()->json(['message' => 'Berhasil Menghapus Kota'], 201);
     }
 
 
-    public function hapusKota($id){
-        $kota = Kota::find($id);
-        $kota->delete();
-        return response()->json(['message'=>'succes delete'], 200);
+    public function hapusKota($id)
+    {
+
+            $kota = Kota::find($id);
+
+            if (!$kota) {
+                return response()->json(['message' => 'Kota Tidak Ditemukan'], 404);
+            }
+
+            $kota->delete();
+            return response()->json(['message' => 'Berhasil Menghapus Kota'], 200);
+  
     }
 }
